[Agent] Extract mixins for turn manager tests
